% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ksrlive-internal.R
\docType{data}
\name{data_kin}
\alias{data_kin}
\title{Time course data for phosphorylation sites}
\format{\preformatted{'data.frame':	84 obs. of  93 variables:
 $ Gene.Name                               : chr  "Bcl2l11" "As160" "Pfkfb2" "Pfkfb2" ...
 $ Phosphorylated.Amino.Acid               : chr  "S" "T" "S" "S" ...
 $ IPI.3.68.Position.in.Protein            : int  83 649 486 469 259 39 348 1248 707 801 ...
 $ IPI                                     : chr  "IPI00119799" "IPI00226750" "IPI00137533" "IPI00137533" ...
 $ Site.Localization.Class                 : chr  "Class I" "Class I" "Class I" "Class I" ...
 $ Localization.Probability                : num  0.998 1 1 1 0.996 ...
 $ Sequence.Window                         : chr  "IFVRRSSLLSRSS" "FRRRAHTFSHPPS" "RRPRNYSVGSRPL" "VRMRRNSFTPLSS" ...
 $ X15sec.Starved..log2..Median            : num  -0.54 1.95 1.1 -0.61 -0.02 0.52 0.09 -0.11 0.44 0.01 ...
 $ X30sec.Starved..log2..Median            : num  -0.44 2.47 1.96 2.35 0.01 1.26 0.25 -0.14 0.84 0.61 ...
 $ X1min.Starved..log2..Median             : num  -1 3.77 2.51 2.1 0 1.5 0.24 0.46 1.85 0.61 ...
 $ X2min.Starved..log2..Median             : num  -1.08 3.46 2.33 2.83 -0.02 1.44 0.17 0.34 3.49 0.28 ...
 $ X5min.Starved..log2..Median             : num  0.37 2.6 2.64 2.98 -0.01 2.13 -0.05 0.55 4.12 0.68 ...
 $ X10min.Starved..log2..Median            : num  0.75 2.79 2.97 2.37 0.2 2.06 -0.01 1.13 3.85 1.39 ...
 $ X20min.Starved..log2..Median            : num  1.33 1.82 3 3.16 0.09 2.1 -0.05 0.9 3.46 0.83 ...
 $ X60min.Starved..log2..Median            : num  0.36 2.73 3.08 2.33 0.1 2.56 0.08 1.31 3.23 1.1 ...
 $ LY.Exps.Insulin.Starved.Median..log2.   : num  0.21 1.96 2.17 2.2 0.15 1.76 -0.14 0.6 2.3 1.23 ...
 $ LY.Exps.Insulin.LY.Insulin.Median..log2.: num  -0.47 -0.6 -0.63 -0.92 -0.3 -1.14 0.15 -0.49 -0.64 0.24 ...
 $ MK.Exps.Insulin.Starved.Median..log2.   : num  -0.02 3.23 2.13 0.74 0.04 1.26 -0.26 0.67 2.77 1.6 ...
 $ MK.Exps.Insulin.MK.Insulin.Median..log2.: num  0.26 -5.41 -0.54 0.84 0.01 -1.36 0.03 -0.33 -0.67 -0.18 ...
 $ Cluster                                 : chr  "D" "A" "B" "B" ...
 $ Akt.Positive.Training.Set               : int  NA 0 0 0 NA NA NA NA NA NA ...
 $ mTOR.Positive.Training.Set              : int  NA NA NA NA NA NA NA NA NA NA ...
 $ PKA.Positive.Training.Set               : int  NA NA NA NA NA NA NA NA NA NA ...
 $ LY.Inhibition.....Median                : num  28 34.1 35.4 47.2 18.6 54.5 9.9 28.9 35.9 -18.1 ...
 $ MK.Inhibition.....Median                : num  16.4 97.6 31 -152.8 -0.5 ...
 $ Mass.Error..ppm.                        : num  -0.3542 0.0223 0.2916 0.3895 -0.4126 ...
 $ PEP                                     : num  3.33e-03 2.78e-14 8.59e-14 8.34e-66 6.01e-32 ...
 $ Search.Score                            : num  129 152 165 207 162 ...
 $ Akt.Prediction.Score                    : num  11.5 92.7 94.4 88.1 16.3 76.1 29.6 47.5 80.5 27.5 ...
 $ Akt.Delta.Score                         : num  -64 74.5 78.7 45 -0.5 54.8 14.6 28.4 38.8 7.8 ...
 $ Akt.Pareto.Ranking                      : num  NaN 3 1 12 NaN 19 89 56 19 NaN ...
 $ mTOR.Prediction.Score                   : num  10.9 18.2 15.7 18.5 5.3 21.3 3.5 8.3 41.7 8 ...
 $ mTOR.Delta.Score                        : num  -64.6 -74.5 -78.7 -69.6 -11.5 -54.8 -26.1 -39.2 -38.8 -19.4 ...
 $ mTOR.Pareto.Ranking                     : num  NaN NaN NaN NaN NaN NaN NaN NaN 66 NaN ...
 $ PKA.Prediction.Score                    : num  75.5 3 8.4 43.1 16.8 8.4 15 19 12.3 19.6 ...
 $ PKA.Delta.Score                         : num  64 -89.7 -85.9 -45 0.5 -67.7 -14.6 -28.4 -68.2 -7.8 ...
 $ PKA.Pareto.Ranking                      : num  17 NaN NaN 73 NaN NaN NaN NaN NaN NaN ...
 $ Uniprot.ID                              : chr  "O54918-1;O54918;Q542N5" "Q8BYJ6-2;Q8BYJ6;Q8BYJ6-1;Q8BYJ6-4;Q8BYJ6-5;Q8BYJ6-3;Q05BK2" "P70265;Q6GTL7;Q8CDS6;B2Z892;B0FLL3" "P70265;Q6GTL7;Q8CDS6;B2Z892;B0FLL3" ...
 $ Proteins.IPI.IDs                        : chr  "IPI00119799" "IPI00226750;IPI00761281;IPI00675197;IPI00928118;IPI00874756;IPI00928545" "IPI00137533;IPI00556828;IPI00938511;IPI00885515" "IPI00137533;IPI00556828;IPI00938511;IPI00885515" ...
 $ Leading.Proteins                        : chr  "IPI00119799" "IPI00226750" "IPI00137533" "IPI00137533" ...
 $ Modified.Sequence                       : chr  "_RSS(ph)IISR_" "_AHT(ph)FSHPPSSSR_" "_NYS(ph)VGSRPIKPISPIR_" "_RNS(ph)FTPISSSNTIR_" ...
 $ TC.Exp1.15sec..log2.                    : num  -0.24 1.95 1.1 1.14 -0.02 1.88 0.32 0.05 0.08 0.59 ...
 $ TC.Exp1.30sec..log2.                    : num  -0.44 2.33 2.42 2.35 0.01 1.76 0.28 0.03 0.89 0.61 ...
 $ TC.Exp1.1min..log2.                     : num  -0.87 2.05 3.06 2.47 0 1.96 0.42 0.22 1.85 1.03 ...
 $ TC.Exp1.2min..log2.                     : num  -0.96 3.02 3.59 2.83 -0.02 2.01 0.47 0.34 3.49 1.54 ...
 $ TC.Exp1.5min..log2.                     : num  0.84 2.6 3.25 2.98 -0.01 2.13 0.14 0.55 4.33 2.09 ...
 $ TC.Exp1.10min..log2.                    : num  1.22 0.73 3.21 3 -0.04 1.85 0.2 0.77 3.9 1.68 ...
 $ TC.Exp1.20min..log2.                    : num  2.07 1.82 3 3.16 0.12 2.1 0.16 0.62 3.49 1.59 ...
 $ TC.Exp1.60min..log2.                    : num  1.26 2.73 3.08 3.3 0.06 2.65 -0.24 0.62 3.81 2.32 ...
 $ TC.Exp2.15sec..log2.                    : num  -0.6 1.5 0.94 -1.65 -0.11 -0.62 0 NaN 0.47 0.01 ...
 $ TC.Exp2.30sec..log2.                    : num  -0.83 2.47 1.96 -0.83 -0.19 -0.11 0.08 NaN 0.84 0.88 ...
 $ TC.Exp2.1min..log2.                     : num  NaN 3.77 2.42 1.85 -0.58 NaN -0.21 NaN 2.18 0.61 ...
 $ TC.Exp2.2min..log2.                     : num  NaN 3.46 1.55 2.18 -0.14 0.37 -0.09 -0.76 3.26 0.28 ...
 $ TC.Exp2.5min..log2.                     : num  NaN 2.47 2.43 2.45 -0.07 1.23 -0.05 0.14 3.79 0.68 ...
 $ TC.Exp2.10min..log2.                    : num  NaN 2.79 2.42 2.16 0.2 NaN -0.01 NaN 3.85 0.68 ...
 $ TC.Exp2.20min..log2.                    : num  0.39 1.28 2.78 2.27 -0.02 1.16 -0.1 NaN 3.46 0.83 ...
 $ TC.Exp2.60min..log2.                    : num  0.2 1.52 2.66 2.33 0.1 0.98 0.08 NaN 3.08 1.1 ...
 $ TC.Exp3.15sec..log2.                    : num  -0.54 3.46 1.2 -0.61 0.06 0.52 0.09 -0.27 0.44 -0.97 ...
 $ TC.Exp3.30sec..log2.                    : num  -0.33 3.8 1.58 2.76 0.19 1.26 0.25 -0.31 0.8 -0.41 ...
 $ TC.Exp3.1min..log2.                     : num  -1.13 4.3 2.51 2.1 0.12 1.04 0.24 0.7 1.71 0.48 ...
 $ TC.Exp3.2min..log2.                     : num  -1.2 3.79 2.33 3.29 0.14 1.44 0.17 0.46 3.6 0.06 ...
 $ TC.Exp3.5min..log2.                     : num  -0.11 3.27 2.64 3.33 0.01 2.21 -0.2 1 4.12 0.21 ...
 $ TC.Exp3.10min..log2.                    : num  0.28 3.46 2.97 2.37 0.22 2.26 -0.07 1.5 3.34 1.39 ...
 $ TC.Exp3.20min..log2.                    : num  1.33 2.7 3.59 3.23 0.09 2.74 -0.05 1.18 3.18 0.09 ...
 $ TC.Exp3.60min..log2.                    : num  0.36 3.22 3.08 2.03 0.25 2.56 0.14 1.99 3.23 0.47 ...
 $ LY.Exp1.Insulin.Starved..log2.          : num  NaN NaN 2.58 NaN 0.04 1.68 0.01 0.6 2.64 NaN ...
 $ LY.Exp1.Insulin.LY.Insulin..log2.       : num  NaN NaN 0.22 NaN 0.03 -0.98 -0.65 -0.49 -0.93 NaN ...
 $ LY.Exp2.Insulin.Starved..log2.          : num  0.16 1.96 1.86 2.44 0.17 2.33 -0.14 2.37 2.3 1.44 ...
 $ LY.Exp2.Insulin.LY.Insulin..log2.       : num  -0.43 -0.6 -0.63 -0.98 -0.44 -2.05 0.15 -2.1 -0.64 0.13 ...
 $ LY.Exp3.Insulin.Starved..log2.          : num  0.27 NaN 2.17 1.96 0.15 1.76 -0.22 -0.01 2.16 1.03 ...
 $ LY.Exp3.Insulin.LY.Insulin..log2.       : num  -0.52 NaN -1.13 -0.87 -0.3 -1.14 0.21 0.28 -0.52 0.34 ...
 $ MK.Exp1.Insulin.Starved..log2.          : num  NaN NaN 1.82 NaN 0.04 1.53 -0.26 0.67 2.44 1.7 ...
 $ MK.Exp1.Insulin.MK.Insulin..log2.       : num  NaN NaN -0.37 NaN 0.04 -1.36 0.01 -0.28 -0.4 -0.15 ...
 $ MK.Exp2.Insulin.Starved..log2.          : num  0.03 3.23 2.13 2.07 0.04 1.26 -0.15 0.75 2.77 1.6 ...
 $ MK.Exp2.Insulin.MK.Insulin..log2.       : num  0.21 -5.41 -0.54 -0.42 0.01 -1.47 0.03 -0.65 -0.67 -0.18 ...
 $ MK.Exp3.Insulin.Starved..log2.          : num  -0.07 NaN 2.33 -0.59 0 0.86 -0.43 0.43 4 1.24 ...
 $ MK.Exp3.Insulin.MK.Insulin..log2.       : num  0.31 NaN -0.77 2.11 -0.22 -1.35 0.45 -0.33 -0.71 -0.42 ...
 $ Uniprot                                 : chr  "O54918" "Q8BYJ6" "P70265" "P70265" ...
 $ SEQ                                     :List of 84
  ..$ : chr "MAKQPSDVSSECDREGGQLQPAERPPQLRPGAPTSLQTEPQGNPDGEGDRCPHGSPQGPLAPPASPGPFATRSPLFIFVRRSSLLSRSSSGYFSFDTDRSPAPMSCDKSTQTPSPPCQAFNHYLSAM"| __truncated__
  ..$ : chr "MESPSCIQDEPFPHPLEPEPSAPAQPGATKPGDKRFRLWYVGGSCLDRRTTLPMLPWLMAEIRRRSQKPDAGGCGAPAAREVILVLSAPFLRCVPAPGAGVGGGAGSGAVQPNTGVFIFEHKAQHIS"| __truncated__
  ..$ : chr "MSENSTFSTEDSCNSSYKPHASNLRRAGKTCSWASYMTNSPTLIVMIGLPARGKTYVSKKLTRYLNWIGVPTKVFNLGVYRREAVKSYQSYDFFRHDNEEAMKIRKQCALVALEDVKAYFTEESGQI"| __truncated__
  ..$ : chr "MSENSTFSTEDSCNSSYKPHASNLRRAGKTCSWASYMTNSPTLIVMIGLPARGKTYVSKKLTRYLNWIGVPTKVFNLGVYRREAVKSYQSYDFFRHDNEEAMKIRKQCALVALEDVKAYFTEESGQI"| __truncated__
  ..$ : chr "MEHIQGAWKTISNGFGLKDAVFDGSSCISPTIVQQFGYQRRASDDGKLTDSSKTSNTIRVFLPNKQRTVVNVRNGMSLHDCLMKALKVRGLQPECCAVFRLLQEHKGKKARLDWNTDAASLIGEELQ"| __truncated__
  ..$ : chr "MSTRSVSSSSYRRMFGGSGTSSRPSSNRSYVTTSTRTYSLGSALRPSTSRSLYSSSPGGAYVTRSSAVRLRSSVPGVRLLQDSVDFSLADAINTEFKNTRTNEKVELQELNDRFANYIDKVRFLEQQ"| __truncated__
  ..$ : chr "MAALSGGGGSSSGGGGGGGGGGGGGDGGGGAEQGQALFNGDMEPEAGAGAAASSAADPAIPEEVWNIKQMIKLTQEHIEALLDKFGGEHNPPSIYLEAYEEYTSKLDALQQREQQLLESLVFQTPTD"| __truncated__
  ..$ : chr "MAGVATPCANGCGPGAPSEAEVLHLCRSLEVGTVMTLFYSKKSQRPERKTFQVKLETRQITWSRGADKIEGSIDIREIKEIRPGKTSRDFDRYQEDPAFRPDQSHCFVILYGMEFRLKTLSLQATSE"| __truncated__
  ..$ : chr "MMLRWSGVWGFHPPRIFPSLLVVVALVGLLPVLRSHGLQHSPTASTIRGSEPPRERSIGDVTTAPSEPLHRPDDHNLTNLIIEHGGKPSRKAFPVLDIDYPHVRTPFEISLWILLACLMKIGFHVIP"| __truncated__
  ..$ : chr "MMLRWSGVWGFHPPRIFPSLLVVVALVGLLPVLRSHGLQHSPTASTIRGSEPPRERSIGDVTTAPSEPLHRPDDHNLTNLIIEHGGKPSRKAFPVLDIDYPHVRTPFEISLWILLACLMKIGFHVIP"| __truncated__
  ..$ : chr "MGNLKSVGQEPGPPCGLGLGLGLGLCGKQGPASPAPEPSQAPAPPSPTRPAPDHSPPLTRPPDGPRFPRVKNWEVGSITYDTLSAQAQQDGPCTSRRCLGSLVFPRKLQSRPTQGPSPTEQLLGQAR"| __truncated__
  ..$ : chr "MATQADLMELDMAMEPDRKAAVSHWQQQSYLDSGIHSGATTTAPSLSGKGNPEEEDVDTSQVLYEWEQGFSQSFTQEQVADIDGQYAMTRAQRVRAAMFPETLDEGMQIPSTQFDAAHPTNVQRLAE"| __truncated__
  ..$ : chr "MLGTGPAVATASAATSSNVSVLQQFASGLKSRNEETRAKAAKELQHYVTMELREMSQEESTRFYDQLNHHIFELVSSSDANERKGGILAIASLIGVEGGNSTRIGRFANYLRNLLPSSDPVVMEMAS"| __truncated__
  ..$ : chr "MAAPSPSGGGGSGGGGGTPGPIGPPASGHPAVSSMQGKRKALKLNFANPPVKSTARFTLNPNTTGVQNPHIERLRTHSIESSGKLKISPEQHWDFTAEDLKDLGEIGRGAYGSVNKMVHKPSGQIMA"| __truncated__
  ..$ : chr "MSNVRVSNGSPSLERMDARQAEHPKPSACRNLFGPVNHEELTRDLEKHCRDMEEASQRKWNFDFQNHKPLEGRYEWQEVERGSLPEFYYRPPRPPKSACKVLAQESQDVSGSRQAVPLIGSQANSED"| __truncated__
  ..$ : chr "MAKPTSKDSGLKEKFKILLGLGTSRPNPRCAEGKQTEFIITSEILRELSGECGLNNRIRMIGQICDVAKTKKLEEHAVEALWKAVSDLLQPERPPEARHAVLTLLKAIVQGQGDRLGVLRALFFKVI"| __truncated__
  ..$ : chr "MAKPTSKDSGLKEKFKILLGLGTSRPNPRCAEGKQTEFIITSEILRELSGECGLNNRIRMIGQICDVAKTKKLEEHAVEALWKAVSDLLQPERPPEARHAVLTLLKAIVQGQGDRLGVLRALFFKVI"| __truncated__
  ..$ : chr "MAKPTSKDSGLKEKFKILLGLGTSRPNPRCAEGKQTEFIITSEILRELSGECGLNNRIRMIGQICDVAKTKKLEEHAVEALWKAVSDLLQPERPPEARHAVLTLLKAIVQGQGDRLGVLRALFFKVI"| __truncated__
  ..$ : chr "MSGRPRTTSFAESCKPVQQPSAFGSMKVSRDKDGSKVTTVVATPGQGPDRPQEVSYTDTKVIGNGSFGVVYQAKLCDSGELVAIKKVLQDKRFKNRELQIMRKLDHCNIVRLRYFFYSSGEKKDEVY"| __truncated__
  ..$ : chr "MASGADSKGDDLSTAILKQKNRPNRLIVDEAINEDNSVVSLSQPKMDELQLFRGDTVLLKGKKRREAVCIVLSDDTCSDEKIRMNRVVRNNLRVRLGDVISIQPCPDVKYGKRIHVLPIDDTVEGIT"| __truncated__
  ..$ : chr "MKLNISFPATGCQKLIEVDDERKLRTFYEKRMATEVAADALGEEWKGYVVRISGGNDKQGFPMKQGVLTHGRVRLLLSKGHSCYRPRRTGERKRKSVRGCIVDANLSVLNLVIVKKGEKDIPGLTDT"| __truncated__
  ..$ : chr "MASLSLAPVNIFKAGADEERAETARLSSFIGAIAIGDLVKSTLGPKGMDKILLSSGRDAALMVTNDGATILKNIGVDNPAAKVLVDMSRVQDDEVGDGTTSVTVLAAELLREAESLIAKKIHPQTII"| __truncated__
  ..$ : chr "MCNTNMSVSTEGAASTSQIPASEQETLVRPKPLLLKLLKSVGAQNDTYTMKEIIFYIGQYIMTKRLYDEKQQHIVYCSNDLLGDVFGVPSFSVKEHRKIYAMIYRNLVAVSQQDSGTSLSESRRQPE"| __truncated__
  ..$ : chr "MCNTNMSVSTEGAASTSQIPASEQETLVRPKPLLLKLLKSVGAQNDTYTMKEIIFYIGQYIMTKRLYDEKQQHIVYCSNDLLGDVFGVPSFSVKEHRKIYAMIYRNLVAVSQQDSGTSLSESRRQPE"| __truncated__
  ..$ : chr "MTTTLVSATIFDLSEVLCKGNKMLNYSTPSAGGCLLDRKAVGTPAGGGFPRRHSVTLPSSKFHQNQLLSSLKGEPAPSLSSRDSRFRDRSFSEGGERLLPTQKQPGSGQVNSSRYKTELCRPFEENG"| __truncated__
  ..$ : chr "MSAGSSCSQTPSRAIPTRRVALGDGVQLPPGDYSTTPGGTLFSTTPGGTRIIYDRKFLMECRNSPVAKTPPKDLPAIPGVTSPTSDEPPMQASQSQLPSSPEDKRAGGEESQFEMDI"
  ..$ : chr "MSAGSSCSQTPSRAIPTRRVALGDGVQLPPGDYSTTPGGTLFSTTPGGTRIIYDRKFLMECRNSPVAKTPPKDLPAIPGVTSPTSDEPPMQASQSQLPSSPEDKRAGGEESQFEMDI"
  ..$ : chr "MSAGSSCSQTPSRAIPTRRVALGDGVQLPPGDYSTTPGGTLFSTTPGGTRIIYDRKFLMECRNSPVAKTPPKDLPAIPGVTSPTSDEPPMQASQSQLPSSPEDKRAGGEESQFEMDI"
  ..$ : chr "MMNNSNYSDASGLGLVDEADEMPSTEKDLAEDAPWKKIQQNTFTRWCNEHLKCVGKRLTDLQRDLSDGLRLIALLEVLSQKRMYRKFHPRPNFRQMKLENVSVALEFLEREHIKLVSIDSKAIVDGN"| __truncated__
  ..$ : chr "MAEYGTLLQDLTNNITLEDLEQLKSACKEDIPSEKSEEITTGSAWFSFLESHNKLDKDNLSYIEHIFEISRRPDLLTMVVDYRTRVLKISEEEELDTKLTRIPSAKKYKDIIRQPSEEEIIKLAPPP"| __truncated__
  ..$ : chr "MDIENEQTLNVNPTDPDNLSDSLFSGDEENAGTEEIKNEINGNWISASTINEARINAKAKRRLRKNSSRDSGRGDSVSDNGSEAVRSGVAVPTSPKGRLLDRRSRSGKGRGLPKKGGAGGKGVWGTP"| __truncated__
  ..$ : chr "MGTPKQPSLAPAHALGLRKSDPGIRSLGSDAGGRRWRPAAQSMFQIPEFEPSEQEDASATDRGLGPSLTEDQPGPYLAPGLLGSNIHQQGRAATNSHHGGAGAMETRSRHSSYPAGTEEDEGMEEEL"| __truncated__
  ..$ : chr "MGTPKQPSLAPAHALGLRKSDPGIRSLGSDAGGRRWRPAAQSMFQIPEFEPSEQEDASATDRGLGPSLTEDQPGPYLAPGLLGSNIHQQGRAATNSHHGGAGAMETRSRHSSYPAGTEEDEGMEEEL"| __truncated__
  ..$ : chr "MGTPKQPSLAPAHALGLRKSDPGIRSLGSDAGGRRWRPAAQSMFQIPEFEPSEQEDASATDRGLGPSLTEDQPGPYLAPGLLGSNIHQQGRAATNSHHGGAGAMETRSRHSSYPAGTEEDEGMEEEL"| __truncated__
  ..$ : chr "MVMADGPRHLQRGPVRVGFYDIEGTLGKGNFAVVKLGRHRTTKTEVAIKIIDKSQLDAVNLEKIYREVQIMKMLDHPHIIKLYQVMETKSMLYLVTEYAKNGEIFDYLANHGRLNESEARRKFWQIL"| __truncated__
  ..$ : chr "MSETSSHDSFYDSLSDVQEEGKSADFFPGLSAFLSQEEINKSLDLARRAIDSSETEDFDSEKEISQIFSKSPISLCETPSHEEPKSGKQTSSERPQDSRRAPVQPLTGDQAERITSPGSKRKPGVSP"| __truncated__
  ..$ : chr "MASGRPEELWEAVVGAAERFQARTGTELVLLTAAPPPPPRPGPCAYAAHGRGALAEAARRCLHDIAQAHRAATATRPPGPPPAPQPPSPAPSPPPRPALAREDEEEDEDEPTETETSGERLGGSDNG"| __truncated__
  ..$ : chr "MAMDWLGSIVSINCGDSLGVYQGRVSAVDQVSQTISLTRPFHNGVKCLVPEVTFRAGDITELKILEIPGPGDNQQVGDLHQTELGPSGVGYQMSISQNGTGKVVKKPASSSSAPQSIPKRTDVKSQD"| __truncated__
  ..$ : chr "MADLANEEKPAVAPSVFVFQKDKGQKRSAGSSSPEAGEDSDHEDGNYCPPVKRERTSSLTHSEEKSSGFRLKPPTLIHGQAPSAGLPSQKPREQQRGVLRPAVLQAPQPKVLSQTVPSSGTNGVSMP"| __truncated__
  ..$ : chr "MMAGAEAPQPQKRYYRQRAHSNPMADHTLRYPVKPEEMDWSELYPEFFAPLIQNKSHDDPKDEKEKHSGAQVEFADIGCGYGGLLVALSPLFPDTLILGLEIRVKVSDYVQDRIRALRAAPGGGFQN"| __truncated__
  ..$ : chr "MAELQEVQITEEKPLLPGQTPETAKEAELAARILLDQGQTHSVETPYGSVTFTVYGTPKPKRPAIFTYHDVGLNYKSCFQPLFRFGDMQEIIQNFVRVHVDAPGMEEGAPVFPLGYQYPSLDQLADM"| __truncated__
  ..$ : chr "MATDDKSSPTLDSANDLPRSPASPSHLTHFKPLTPDQDEPPFKSAYSSFVNLFRFNKERGEGGQGEQQSPSSSWASPQIPSRTQSVRSPVPYKKQLNEELHRRSSVLDSRRKAEPACGGHDPRTAVQ"| __truncated__
  ..$ : chr "MSSEPPPPPLQPPTHQTSVGLLDTPRTRDRSPSPLRGNVVPSPLPTRRTRTFSATVRASQGPVYKGVCKCFCRSKGHGFITPADGGPDIFLHISDVEGEYVPVEGDEVTYKMCSIPPKNEKLQAVEV"| __truncated__
  ..$ : chr "MPLFGSIFSPKKTPPRKSASLSNLHSLDRSTRELELGLDYGTPTMNLAGQSLKFENGQWVADSVISGGVDRRETQRLRKRNQQLEEENNLLRLKVDILLDMLSETTAESHLKDKELDELKVTNRRRK"| __truncated__
  ..$ : chr "MADEDLIFCLEGVDGGRCSRAGHNADSDTDSDDDEGYFICPITDDHMSNQNVSSKVQSYYSNLTKTECGSTGSPASSFHFKEAWKHAIEKAKHMPDPWAEFHLEDIATEHATRHRYNAVTGEWLKDE"| __truncated__
  ..$ : chr "MEKATVPAAAEGEGSPPAAAAVAAPPAAAAAEVGGGARPASSPRGMVRVCDLLLKKKPPQQQQQQQPPHHKAKRNRTCRPPSSSESSSDSDNSGGGGGGGGGGGGGTSSNNSEEEEDDDDEEEEVSE"| __truncated__
  ..$ : chr "MEKATVPAAAEGEGSPPAAAAVAAPPAAAAAEVGGGARPASSPRGMVRVCDLLLKKKPPQQQQQQQPPHHKAKRNRTCRPPSSSESSSDSDNSGGGGGGGGGGGGGTSSNNSEEEEDDDDEEEEVSE"| __truncated__
  ..$ : chr "MEPGRGGVETVGKFEFSRKDLIGHGAFAVVFKGRHREKHDLEVAVKCINKKNLAKSQTLLGKEIKILKELKHENIVALYDFQEMANSVYLVMEYCNGGDLADYLHTMRTLSEDTVRLFLQQIAGAMR"| __truncated__
  ..$ : chr "MRRRRRRDGFYLAPDFRHREAEDMAGVFDIDLDQPEDAGSEDELEEGGQLNESMDHGGVGPYELGMEHCEKFEISETSVNRGPEKIRPECFELLRVLGKGGYGKVFQVRKVTGANTGKIFAMKVLKK"| __truncated__
  ..$ : chr "MRRRRRRDGFYLAPDFRHREAEDMAGVFDIDLDQPEDAGSEDELEEGGQLNESMDHGGVGPYELGMEHCEKFEISETSVNRGPEKIRPECFELLRVLGKGGYGKVFQVRKVTGANTGKIFAMKVLKK"| __truncated__
  ..$ : chr "MASPPDTDGFSDVRKVGYLRKPKSMHKRFFVLRAASEAGGPARLEYYENEKKWRHKSSAPKRSIPLESCFNINKRADSKNKHLVALYTRDEHFAIAADSEAEQDSWYQALLQLHNRAKAHHDGAGGG"| __truncated__
  ..$ : chr "MAEVGEDSGARALLALRSAPCSPVLCAAAAAAAFPATTSPPPPAQPPPGPPALPAEPGPGPVPSTVATATTTAPALVAAAAASVRQSPGPALARLEGREFEFLMRQPSVTIGRNSSQGSVDLSMGLS"| __truncated__
  ..$ : chr "MAEVGEDSGARALLALRSAPCSPVLCAAAAAAAFPATTSPPPPAQPPPGPPALPAEPGPGPVPSTVATATTTAPALVAAAAASVRQSPGPALARLEGREFEFLMRQPSVTIGRNSSQGSVDLSMGLS"| __truncated__
  ..$ : chr "MLGTGPAVATASAATSSNVSVLQQFASGLKSRNEETRAKAAKELQHYVTMELREMSQEESTRFYDQLNHHIFELVSSSDANERKGGILAIASLIGVEGGNSTRIGRFANYLRNLLPSSDPVVMEMAS"| __truncated__
  ..$ : chr "MLGTGPAVATASAATSSNVSVLQQFASGLKSRNEETRAKAAKELQHYVTMELREMSQEESTRFYDQLNHHIFELVSSSDANERKGGILAIASLIGVEGGNSTRIGRFANYLRNLLPSSDPVVMEMAS"| __truncated__
  ..$ : chr "MLGTGPAVATASAATSSNVSVLQQFASGLKSRNEETRAKAAKELQHYVTMELREMSQEESTRFYDQLNHHIFELVSSSDANERKGGILAIASLIGVEGGNSTRIGRFANYLRNLLPSSDPVVMEMAS"| __truncated__
  ..$ : chr "MLGTGPAVATASAATSSNVSVLQQFASGLKSRNEETRAKAAKELQHYVTMELREMSQEESTRFYDQLNHHIFELVSSSDANERKGGILAIASLIGVEGGNSTRIGRFANYLRNLLPSSDPVVMEMAS"| __truncated__
  ..$ : chr "MLGTGPAVATASAATSSNVSVLQQFASGLKSRNEETRAKAAKELQHYVTMELREMSQEESTRFYDQLNHHIFELVSSSDANERKGGILAIASLIGVEGGNSTRIGRFANYLRNLLPSSDPVVMEMAS"| __truncated__
  ..$ : chr "MSSPPEGKLETKAGHPPAVKAGGMRIVQKHPHTGDGKEERDKDDQEWESTSPPKPTVFISGVIARGDKDFPPAAAQVAHQKPHASMDKHVSPRTQHIQQPRK"
  ..$ : chr "MSAGSSCSQTPSRAIPTRRVALGDGVQLPPGDYSTTPGGTLFSTTPGGTRIIYDRKFLMECRNSPVAKTPPKDLPAIPGVTSPTSDEPPMQASQSQLPSSPEDKRAGGEESQFEMDI"
  ..$ : chr "MSAGSSCSQTPSRAIPTRRVALGDGVQLPPGDYSTTPGGTLFSTTPGGTRIIYDRKFLMECRNSPVAKTPPKDLPAIPGVTSPTSDEPPMQASQSQLPSSPEDKRAGGEESQFEMDI"
  ..$ : chr "MSAGSSCSQTPSRAIPTRRVALGDGVQLPPGDYSTTPGGTLFSTTPGGTRIIYDRKFLMECRNSPVAKTPPKDLPAIPGVTSPTSDEPPMQASQSQLPSSPEDKRAGGEESQFEMDI"
  ..$ : chr "MSAGSSCSQTPSRAIPTRRVALGDGVQLPPGDYSTTPGGTLFSTTPGGTRIIYDRKFLMECRNSPVAKTPPKDLPAIPGVTSPTSDEPPMQASQSQLPSSPEDKRAGGEESQFEMDI"
  ..$ : chr "MSAGSSCSQTPSRAIPTRRVALGDGVQLPPGDYSTTPGGTLFSTTPGGTRIIYDRKFLMECRNSPVAKTPPKDLPAIPGVTSPTSDEPPMQASQSQLPSSPEDKRAGGEESQFEMDI"
  ..$ : chr "MISAPDVVAFTKEDEYEEEPYNEPALPEEYSVPLFPYASQGANPWSKLSGAKFSRDFILISEFSEQVGPQPLLTIPNDTKVFGTFDLNYFSLRIMSVDYQASFVGHPPGSAYPKLNFVEDSKVVLGD"| __truncated__
  ..$ : chr "MATQVEPLLPAGAPLLQAEEHGLARKKPAPDAQAESGPGDGGGEPDGGVRRPRPACARPGRDGAERESPRPPAAAEAPAGSDGEDGGRRDFVEAPPPKVNPWTKHAPPPAAVNGQPPPEPSAPAKVV"| __truncated__
  ..$ : chr "MATQVEPLLPAGAPLLQAEEHGLARKKPAPDAQAESGPGDGGGEPDGGVRRPRPACARPGRDGAERESPRPPAAAEAPAGSDGEDGGRRDFVEAPPPKVNPWTKHAPPPAAVNGQPPPEPSAPAKVV"| __truncated__
  ..$ : chr "MSAKVRLKKLEQLLLDGPWRNDSALSVETLLDVLVCLYTECSHSALRRDKYVAEFLEWAKPFTQLVKDMQLHREDFEIIKVIGRGAFGEVAVVKMKNTERIYAMKILNKWEMLKRAETACFREERDV"| __truncated__
  ..$ : chr "MNLASQSGEAGAGQLLFANFNQDNTSLAVGSKSGYKFFSLSSVDKLEQIYECTDTEDVCIVERLFSSSLVAIVSLKAPRKLKVCHFKKGTEICNYSYSNTILAVKLNRQRLIVCLEESLYIHNIRDM"| __truncated__
  ..$ : chr "MEELIVELRLFLELLDHEYLTSTVREKKAVLTNILLRLQSSKGFEVKDHAQKAEANNLPAPPQMPLPEIPQPWLPPDSGPPPLPTSSLPEGYYEEAVPLSPGKAPEYITSNYDSDAMSSSYESYDEE"| __truncated__
  ..$ : chr "MEELIVELRLFLELLDHEYLTSTVREKKAVLTNILLRLQSSKGFEVKDHAQKAEANNLPAPPQMPLPEIPQPWLPPDSGPPPLPTSSLPEGYYEEAVPLSPGKAPEYITSNYDSDAMSSSYESYDEE"| __truncated__
  ..$ : chr "MFRYESLEDCPLDEDEDAFQGLGEEDEEIDQFNDDTFGSGAVDDDWQEAHERLAELEEKLPVAADEQTGNGERDEMDLLGDHEENLAERLSKMVIENELEDPAIMRAVQTRPVLQPQPGSLNSSIWD"| __truncated__
  ..$ : chr "MFRYESLEDCPLDEDEDAFQGLGEEDEEIDQFNDDTFGSGAVDDDWQEAHERLAELEEKLPVAADEQTGNGERDEMDLLGDHEENLAERLSKMVIENELEDPAIMRAVQTRPVLQPQPGSLNSSIWD"| __truncated__
  ..$ : chr "METGPAPLVAPPRRHGAPAAPSPPPRGSRAGSHLVVEPGPPVTTATSAPVELVAPGEARPACVPGSSQTSASTPTTATSSTVVMLTLEASPEAAKTQEFPAPAAETGAETSVALALGTDTQKTEEVR"| __truncated__
  ..$ : chr "MDAGFFRGTSAEQDNRFSNKQKKLLKQLKFAECLEKKVDMSKVNLEVIKPWITKRVTEILGFEDDVVIEFIFNQLEVKNPDSKMMQINLTGFLNGKNAREFMGELWPLLLSAQENIAGIPSAFLELK"| __truncated__
  ..$ : chr "MDAGFFRGTSAEQDNRFSNKQKKLLKQLKFAECLEKKVDMSKVNLEVIKPWITKRVTEILGFEDDVVIEFIFNQLEVKNPDSKMMQINLTGFLNGKNAREFMGELWPLLLSAQENIAGIPSAFLELK"| __truncated__
  ..$ : chr "MESEMLQSPLMGLGEEDEADLTDWNLPLAFMKKRHCEKIEGSKSLAQSWRMKDRMKTVSVALVLCLNVGVDPPDVVKTTPCARLECWIDPLSMGPQKALETIGANLQKQYENWQPRARYKQSLDPTV"| __truncated__
  ..$ : chr "MESEMLQSPLMGLGEEDEADLTDWNLPLAFMKKRHCEKIEGSKSLAQSWRMKDRMKTVSVALVLCLNVGVDPPDVVKTTPCARLECWIDPLSMGPQKALETIGANLQKQYENWQPRARYKQSLDPTV"| __truncated__
  ..$ : chr "MEEGSSGGSGSSDSNAGGSGGVQQRELERMAEVLVTGEQLRLRLHEEKVIKDRRHHLKTYPNCFVAKELIDWLIEHKEASDRETAIKLMQKLADRGIIHHVCDEHKEFKDVKLFYRFRKDDGTFALD"| __truncated__
  ..$ : chr "MASGRPEELWEAVVGAAERFQARTGTELVLLTAAPPPPPRPGPCAYAAHGRGALAEAARRCLHDIAQAHRAATATRPPGPPPAPQPPSPAPSPPPRPALAREDEEEDEDEPTETETSGERLGGSDNG"| __truncated__
  ..$ : chr "MKLLENSSFEAINSQLTVETGDAHIIGRIESYSCKMAGDDKHMFKQFCQEGQPHVLEALSPPQTSGLSPSRLSKSQGGEDESPLSDKCSRKTLFYLIATLNESFRPDYDFSTARSHEFSREPSLRWV"| __truncated__
  ..$ : chr "MAAATGAGRLRRAASALLLRSPRLPARELSAPARLYHKKVVDHYENPRNVGSLDKTSKNVGTGLVGAPACGDVMKLQIQVDEKGKIVDARFKTFGCGSAIASSSLATEWVKGKTVEEALTIKNTDIA"| __truncated__
  ..$ : chr "MSDGAAEKQSGTPGFLTPPAPVPKNGSSSDSSVGEKLGATVADSGVGRTEEYRRRRHTMDKDSRGAAATTTPTEHRFFRRSVICDSNATALELPGLPLSIPQPSVPAVVPQSAPPEPHREETLTATV"| __truncated__
  ..$ : chr "MYNGIGLPTPRGSGTNGYVQRNLSLVRGRRGERPDYKGEEELRHLEAALVKRPNPDILDHERKRRVELRCLELEEMMEEQGYEEQQIQEKVATFRLMLLEKDVNPGAKEETPGQRPVVTETHQLAEL"| __truncated__
 $ Motif                                   : chr  "FIFVRRSSLLSRSSS" "QFRRRAHTFSHPPSS" "IRRPRNYSVGSRPLK" "PVRMRRNSFTPLSSS" ...
 $ Position                                : num  83 649 486 469 259 ...
 $ 0_scaled                                : num  0.4481 0 0 0.1618 0.0909 ...
 $ 15s_scaled                              : num  0.224 0.517 0.357 0 0 ...
 $ 30s_scaled                              : num  0.266 0.655 0.636 0.785 0.136 ...
 $ 1min_scaled                             : num  0.0332 1 0.8149 0.7188 0.0909 ...
 $ 2min_scaled                             : num  0 0.918 0.756 0.912 0 ...
 $ 5min_scaled                             : num  0.6017 0.6897 0.8571 0.9523 0.0455 ...
 $ 10min_scaled                            : num  0.759 0.74 0.964 0.79 1 ...
 $ 20min_scaled                            : num  1 0.483 0.974 1 0.5 ...
 $ 60min_scaled                            : num  0.598 0.724 1 0.78 0.545 ...
 $ names                                   : chr  "O54918_83" "Q8BYJ6_649" "P70265_486" "P70265_469" ...
 $ Uniprot.human                           : chr  "O43521" "O60343" "O60825" "O60825" ...
 $ Motif.human                             : chr  "FIFMRRSSLLSRSSS" "QFRRRAHTFSHPPSS" "IRRPRNYSVGSRPLK" "PVRMRRNSFTPLSSS" ...
}}
\source{
Humphrey et al., Cell Metabolism, 2013
}
\usage{
data_kin
}
\description{
This dataset contains time course data of phosphorylation sites after insulin
stimulation.
}
\keyword{datasets}

