% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ksrlive-internal.R
\docType{data}
\name{data_list}
\alias{data_list}
\title{List containing time course data for phosphorylation sites}
\format{\preformatted{List of 2
 $ P31749:'data.frame':	15 obs. of  9 variables:
  ..$ 0_scaled    : num [1:15] 0.4481 0 0 0 0.0362 ...
  ..$ 15s_scaled  : num [1:15] 0.22407 0.51724 0.35714 0.00719 0 ...
  ..$ 30s_scaled  : num [1:15] 0.266 0.655 0.636 0.439 0.102 ...
  ..$ 1min_scaled : num [1:15] 0.0332 1 0.8149 0.4388 0.371 ...
  ..$ 2min_scaled : num [1:15] 0 0.918 0.756 0.201 0.52 ...
  ..$ 5min_scaled : num [1:15] 0.602 0.69 0.857 0.489 0.876 ...
  ..$ 10min_scaled: num [1:15] 0.759 0.74 0.964 1 1 ...
  ..$ 20min_scaled: num [1:15] 1 0.483 0.974 0.597 0.925 ...
  ..$ 60min_scaled: num [1:15] 0.598 0.724 1 0.791 0.846 ...
 $ P42345:'data.frame':	28 obs. of  9 variables:
  ..$ 0_scaled    : num [1:28] 0.7765 0.6528 0.3448 0.75 0.0404 ...
  ..$ 15s_scaled  : num [1:28] 0.7095 0.7222 0 0.9621 0.0101 ...
  ..$ 30s_scaled  : num [1:28] 0.7374 0.6667 0.0172 1 0 ...
  ..$ 1min_scaled : num [1:28] 0.866 0.764 0.138 0.765 0.247 ...
  ..$ 2min_scaled : num [1:28] 1 0.736 0.241 0.75 0.535 ...
  ..$ 5min_scaled : num [1:28] 0.246 1 0.534 0.939 0.808 ...
  ..$ 10min_scaled: num [1:28] 0.33 0.556 0.81 0.424 1 ...
  ..$ 20min_scaled: num [1:28] 0.184 0.333 0.69 0.5 0.854 ...
  ..$ 60min_scaled: num [1:28] 0 0 1 0 0.626 ...
}}
\source{
Humphrey et al., Cell Metabolism, 2013
}
\usage{
data_list
}
\description{
This dataset contains time course data of phosphorylation sites after insulin
stimulation.
}
\keyword{datasets}

