% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kinclust.R
\name{kinclust}
\alias{kinclust}
\title{Find the lowest pvalue in clustering.}
\usage{
kinclust(data_list, cl = NULL, ...)
}
\arguments{
\item{result}{Object created by pvclust}
}
\value{
result

most_stable

clustering

data
}
\description{
\code{kinclust} returns highest au that still has lowest pvalue
}
\details{
Function most.stable takes an object created by pvclust
 and finds the highest au that has lowest pvalue
}
\examples{
data(mtcars)
data <- t(mtcars)
result <- pvclust(data)
mostab <- most.stable(result)
}

