% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksrlive.R
\docType{package}
\name{ksrlive}
\alias{ksrlive}
\alias{ksrlive-package}
\title{Identify site specific kinase substrate relationships using dynamic data.}
\description{
Using this package you can combine known site specific 
kinase substrate relationships with dynamic experimental data and determine active 
kinases and their substrates.

}
\author{
Westa Domanova
}

