% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ksrlive-internal.R
\name{most.stable}
\alias{most.stable}
\title{Find the lowest pvalue in clustering.}
\usage{
most.stable(result)
}
\arguments{
\item{result}{Object created by pvclust}
}
\value{
numeric of the highest au
}
\description{
\code{most.stable} returns highest au that still has lowest pvalue
}
\details{
Function most.stable takes an object created by pvclust
 and finds the highest au that has lowest pvalue
}
\examples{
data(mtcars)
data <- t(mtcars)
result <- pvclust(data)
mostab <- most.stable(result)
}

