% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ksrlive-internal.R
\name{plotlines}
\alias{plotlines}
\title{Create cluster assignment vector from pvpick object.}
\usage{
plotlines(plotnames, data, clust.class = c(1:length(plotnames)),
  colors = rainbow(length(plotnames)), scaled = TRUE, label = "none")
}
\arguments{
\item{data}{Data used for clustering}

\item{pvclust_pick}{Object created by pvclust::pvpick}
}
\value{
named vector with cluster assignments for all instances that were
clustered
}
\description{
\code{pvclust.clust} returns a named vector with cluster assignments
}
\details{
Function pvclust.clust takes an object created by pvpick from the pvclust
package and transforms it into a named vector with cluster assignments.
}
\examples{
data(mtcars)
data <- t(mtcars)
result <- pvclust(data)
pvclust_pick <- pvpick(result, alpha = 0.95)
clust_class <- pvclust.clust(pvclust_pick, data)
}

