% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ksrlive-internal.R
\name{pvclust.clust}
\alias{pvclust.clust}
\title{Create cluster assignment vector from pvpick object.}
\usage{
pvclust.clust(pvclust_pick, data)
}
\arguments{
\item{pvclust_pick}{Object created by pvclust::pvpick}

\item{data}{Data used for clustering}
}
\value{
named vector with cluster assignments for all instances that were
clustered
}
\description{
\code{pvclust.clust} returns a named vector with cluster assignments
}
\details{
Function pvclust.clust takes an object created by pvpick from the pvclust
package and transforms it into a named vector with cluster assignments.
}
\examples{
data(mtcars)
data <- t(mtcars)
library(pvclust)
result <- pvclust(data)
pvclust_pick <- pvpick(result, alpha = 0.95)
clust_class <- pvclust.clust(pvclust_pick, data)
}

