% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ksrlive.R
\name{clust.expand}
\alias{clust.expand}
\title{Integrate exclusive and complete clustering results}
\usage{
clust.expand(kin_clust, kin_clust_all, thre = 0.95,
  index = 1:length(kin_clust[[1]]), diff_list = NULL)
}
\arguments{
\item{kin_clust}{list of kinase substrate relationships with only exclusive
substrates}

\item{kin_clust_all}{list of all available kinase substrate relationships}

\item{thre}{threshold of au value for clustering, for example au = 0.95
corresponds to a pvalue of 0.05}

\item{index}{numeric vector of indices for which to do the integration}

\item{diff_list}{character vector of names of differentially regulated
substrates}
}
\value{
expand_clust_list integrated clustering assignments and noclust

noclust numeric vector of indices where a cluster could not be
determined
}
\description{
\code{clust.expand} returns a list of clustering assignments
}
\details{
The function clust.expand takes two objects created by kinclust, one
clustered using exclusive substrates and the other one all substrates and
expands the core found by clustering exclusive substrates using a pvalue
threshold. The core sites will be tested for differential regulation if
a list of differentially regulated sites is included (recommended).
}
\examples{
jjj
}

