% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ksrlive.R
\name{kinclust}
\alias{kinclust}
\title{Do pvclust clustering for data in list}
\usage{
kinclust(data_list, cl = NULL, ...)
}
\arguments{
\item{data_list}{list containing time course data for substrates of a kinase
where rows correspond to substrates and a column to the observation at a
time point}

\item{cl}{cluster object created by the parallel package}

\item{...}{arguments used in pvclust}
}
\value{
result list of pvclust objects

most_stable list of aus used for cluster determination

clustering list of named vectors with cluster assignment

data list of data frames used for clustering
}
\description{
\code{kinclust} returns list of clustering results for a given kinase
substrate list
}
\details{
This is the main function of the package. The function kinclust takes a list of kinase substrate relationships,
runs clustering using pvclust on every list and returns a list with
the resulting pvclust objects and cluster assignments.
}
\examples{
data(datalist)
kin_clust <- kinclust(data_list = data_list,
nboot = 1000,
method.dist = "euclidean",
method.hclust = "average")

# using a cluster
library(parallel)
cl <- makeCluster(detectCores())
kin_clust <- kinclust(data_list = data_list,
cl = cl,
nboot = 1000,
method.dist = "euclidean",
method.hclust = "average",
init.rand=TRUE, iseed=21)
stopCluster(cl)
}

