% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ksrlive.R
\name{KSR.list}
\alias{KSR.list}
\title{Create a kinase substrate list from a data frame}
\usage{
KSR.list(db, kinasefamilies = NULL, exclusive = FALSE)
}
\arguments{
\item{db}{data frame of kinase substrate relationships with substrate
identifier in the first column and kinase identifier in the second column.
Data base substrate/kinase identifiers should correspond to data
substrate/kinase identifiers.}

\item{kinasefamilies}{list of kinase identifiers that have to be combined,
one list per kinase family}

\item{exclusive}{logical, if TRUE only substrates exclusive to the kinase
will be included in the list}
}
\value{
list with kinase substrate relationships, with the kinase
identifiers as the list names
}
\description{
\code{KSR.list} returns list of kinase substrate relationships
}
\details{
The function KSR.list creates a list of kinase substrate relationships from
a data frame and can combine kinase families into one list.
}
\examples{
data(phosphonetwork)
test_db <- do.call(paste,
c(phosphonetwork.df[ ,c("SUB_ACC_ID.human", "MODSITE_SEQ.human")],
sep="_"))
test_db <- data.frame(substrate = test_db,
kinase = phosphonetwork.df[ ,"KIN_ACC_ID.human"],
stringsAsFactors = FALSE)
kin_list <- KSR.list(test_db)
# using family lists
fam <- list(akt = c("P31749", "P31751"),
rsk = c("Q15418", "P23443"))
kin.fam <- KSR.list(test_db, kinasefamilies=fam)
kin.fam.exc <- KSR.list(test_db, kinasefamilies=fam, exclusive=TRUE)
}

